requirements.txt
aiogram==3.23.0
Flask==3.1.2
python-dotenv==1.0.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.3.0
sqlalchemy
psycopg2-binary
cryptography
gunicorn
pytest==9.0.2
joblib
pandas
nltk
scikit-learn

pytest-asyncio>=0.24.2
pytest-mock==3.14.0
httpretty==1.1.4
freezegun==1.5.1
factory-boy==3.3.0
faker==24.11.0
asynctest==0.13.0
aioresponses==0.7.6
pytest-cov